App_MuseeAPI/requirements.txt
django==1.11.17
djangorestframework==3.9.0
django-extensions==2.1.4
pillow==5.3.0
